requirements.txt
openai>=1.50.0
mcp>=1.2.0
pyyaml>=6.0
prompt_toolkit>=3.0.43
rich>=13.7.0
rank_bm25>=0.2.2
